djangoBackend/cepv_backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
pandas==3.0.0
numpy==2.2.6
python-dotenv==1.2.1
reportlab==4.4.9
gunicorn==25.0.0
whitenoise==6.11.0
dj-database-url==3.1.0
psycopg2-binary==2.9.11
